python-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
requests==2.32.3
werkzeug==3.0.1
gunicorn==21.2.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchvision==0.20.1
